tests: unit-tests: qatozil: Update tests so that they are run against the app src
