+++ .gitignore
@@ -7,3 +7,6 @@ src/glyphs.h
 *.pyc
 tests/elfs/
 tests/functional/snapshots-tmp/
+tests/unit-tests/qatozil/*.o
+tests/unit-tests/qatozil/*.o.d
+tests/unit-tests/qatozil/qatozil
