--- test/qatozil/Makefile
@@ -1,2 +1,0 @@
-all: qatozil.c
-	gcc -ggdb3 -Wall qatozil.c -o qatozil
