+++ tests/unit-tests/qatozil/test.sh
@@ -1,3 +1,5 @@
+#!/usr/bin/bash
+
 inputs=(
     1
     0
@@ -18,23 +20,6 @@ do
         echo "Testing $item failed: $qatozil received vs $verifier expected"
         exit 1
     fi
-done
-
-shifti=2
-i=0
-while [[ $i -le 10001 ]]
-do
-    qatozil=`./qatozil -shift $shifti $i`
-    verifier=`./verifier.py -shift $shifti $i`
-    echo -n "Testing \"qatozil -shift $shifti $i\" "
-    if [[ $qatozil != $verifier ]]
-    then
-        echo "failed: $qatozil received vs $verifier expected"
-        exit 1
-    else
-        echo "success"
-    fi
-    i=$(($i+1))
 done
 
 echo "All tests completed successfully"
